# agro-api/app/Models/Praga.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Builder;

/**
 * @mixin Builder
 * @OA\Schema(
 *   @OA\Xml(name="Praga")
 * )
 */
class Praga extends Model
{
    //
}