# agro-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/pragas": {
            "get": {
                "tags": [
                    "Pragas"
                ],
                "summary": "Obter lista de pragas",
                "description": "Retorna uma lista de pragas",
                "operationId": "getPragas",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PragaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Praga": {
                "xml": {
                    "name": "Praga"
                }
            },
            "PragaResource": {
                "title": "PragaResource",
                "description": "Project resource",
                "properties": {
                    "data": {
                        "title": "Dados",
                        "description": "Lista de pragas",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Praga"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PragaResource"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ]
}